requirements.txt
anthropic>=0.34.0
chromadb>=0.5.5
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-chroma>=0.1.4
langchain-community>=0.3.0
python-docx>=1.1.0
pypdf>=4.3.0
# sentence-transformers is optional; chromadb's DefaultEmbeddingFunction (ONNX) is used instead
# to avoid torch/numpy version conflicts. Uncomment if you need fine-grained model control:
# sentence-transformers>=3.1.0
pydantic>=2.8.0
python-dotenv>=1.0.0
rich>=13.8.0
streamlit>=1.38.0
pyyaml>=6.0.0
pytest>=8.3.0
pandas>=2.0.0
